respect move of SynapseTable into table subdir
