+++ src/lib/containers/QueryWrapperMenu.tsx
@@ -6,7 +6,7 @@ import { Facets } from './Facets'
 import QueryCount from './QueryCount'
 import QueryWrapper from './QueryWrapper'
 import StackedBarChart from './StackedBarChart'
-import SynapseTable, { SynapseTableProps, unCamelCase } from './SynapseTable'
+import SynapseTable, { SynapseTableProps, unCamelCase } from './table/SynapseTable'
 import CardContainer from './CardContainer'
 import { CommonCardProps } from './CardContainerLogic'
 import { StackedBarChartProps } from './StackedBarChart'
