+++ src/lib/containers/table/SynapseTable.tsx
@@ -860,10 +860,11 @@ export default class SynapseTable extends React.Component<
           ? 'SRC-selected-table-icon'
           : 'SRC-primary-text-color'
         const sortSpanBackgoundClass = `SRC-tableHead SRC-hand-cursor SRC-sortPadding SRC-primary-background-color-hover  ${isSelectedSpanClass}`
+        const displayColumnName: string | undefined = unCamelCase(column.name)
         return (
           <th key={column.name}>
             <div className="SRC-centerContent">
-              <span style={{ whiteSpace: 'nowrap' }}>{column.name}</span>
+              <span style={{ whiteSpace: 'nowrap' }}>{displayColumnName}</span>
               {isFacetSelection &&
                 this.configureFacetDropdown(facets, facetIndex)}
               <span
@@ -1039,7 +1040,19 @@ export default class SynapseTable extends React.Component<
     this.props.executeQueryRequest!(newQueryRequest)
   }
 }
-
+export const unCamelCase = (
+  str: string | undefined
+  ): string | undefined => {
+    // https://stackoverflow.com/questions/4149276/how-to-convert-camelcase-to-camel-case
+    if (!str) { return str }
+    return str
+        // insert a space between lower & upper
+        .replace(/([a-z])([A-Z])/g, '$1 $2')
+        // space before last upper in a sequence followed by lower
+        .replace(/\b([A-Z]+)([A-Z])([a-z])/, '$1 $2$3')
+        // uppercase the first character
+        .replace(/^./, (str:string) => { return str.toUpperCase()})
+}
 type ColumnReference = {
   index: number
   name: string
